Added Powers and Weaknesses Tables
Had previously left off that the user needed to be able to add to the Powers and Weaknesses tables.  This version creates those tables as well as the form to insert into the tables.  Power and Weakness do not have the ability to Delete because the requirements do not state that you need to be able to delete from every table.  The many-to-many tables relationships do allow for deletion.